Tool for comparing firmware contents, some initial analysis
Since the data is encrypted in blocks of 1K, it's possible to get some
insight into the commonality between the firmware images for each MCU.
Taking it further, why not compare blocks across any number of files,
and find where data is same, different, or copied to other locations.

The code is pretty horrible and hacky, cleanup is in-progress.