Cable issue resolved - need a pullup resistor to avoid dropping joystick
The line I've been calling TX goes from the computer/joystick to the MCUs.
It's the white wire on USB and also white on the 6-pin cable in the gimbal.
However, it's actually labeled RX on the joystick/USB board, since it's
receiving data from the gimbal's point of view.

Apparently the weird results I've been getting without the serial adapter were
from extra wire capacitance keeping the line low. This corrupted the joystick
packets and prevented the MCU from recognizing it. Adding a 10k Ohm resistor
between the white wire and 5v solves the issue.